+++ gulpfile.js
@@ -30,11 +30,6 @@ var gulp = require('gulp'),
 gulp.task('vendors', function() {
 
   gulp.src([
-      'bower_components/TimelineJS/build/**/*'
-    ])
-    .pipe(gulp.dest('build/js/TimelineJS'));
-
-  gulp.src([
       'bower_components/bootstrap-sass-official/vendor/assets/fonts/bootstrap/*'
     ])
     .pipe(gulp.dest('build/css/bootstrap'));